Small preparations
Made changes to the dayExpanded for, this will be changed into a timer create form later with a new day expanded to replace it containing buttons to create view statistics ect, Having it all on one page will be too messy.
Added a way to implement dynamic icons depending on the date, all icons still need to be added, this should be in the next commit.